requirements.txt
anyio==4.13.0
asgiref==3.8.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==5.4.0
django-sendgrid-v5==1.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.12
Markdown==3.8
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
payu-websdk==0.1.7
proto-plus==1.27.2
protobuf==6.33.6
psycopg2==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-http-client==3.3.7
razorpay==2.0.1
redis==6.2.0
requests==2.33.1
sendgrid==6.12.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
whitenoise==6.8.2
